Created IDamageable interface
The IDamageable interface can be added where a class is defined as MonoBehaviour, and requires that class to implement a Damage() function and a health variable.
When any weapon (player or enemy) collides with something, it will check for the IDamageable interface. If the object can be damaged, the Damage() function is called, regardless of object type